Moved transform code to transform.ts.
